Refactored code and implemented a watchlist page with temp data
