+++ src/Front-end/ListingsLayout.js
@@ -4,13 +4,9 @@ import { CATEGORIES, CONDITIONS, SORTING, MAXPRICE } from './util';
 import NavBar from './NavBar';
 import { pageTheme } from './util';
 
-const listings = [
-  { id: 1, title: 'CS 121 Textbook', category: 'Books', description: 'So many pages—all so exciting!', condition: 'Like New', price: 15, imageUrl: 'https://via.placeholder.com/150', postDate: '2024-03-28' },
-  { id: 2, title: 'Chair', category: 'Furniture', description: 'Good for sitting in. Excellent for dancing on.', condition: 'Very Good', price: 50, imageUrl: 'https://via.placeholder.com/150', postDate: '2024-03-27' },
-  { id: 3, title: 'MacBook Pro', category: 'Electronics', description: 'Still in its original box. Decided to give up on my CS degree, so I never ended up using this.', condition: 'New', price: 500, imageUrl: 'https://via.placeholder.com/150', postDate: '2024-03-26' },
-];
+const ListingsLayout = (props) => {
+  const {title, listings} = props
 
-const HomePage = () => {
   const [checkedCategories, setCheckedCategories] = useState(CATEGORIES);
   const [checkedConditions, setCheckedConditions] = useState(CONDITIONS);
   const [priceRange, setPriceRange] = useState([0, MAXPRICE]);
@@ -45,7 +41,7 @@ const HomePage = () => {
     }
 
     return filteredListings;
-  }, [checkedCategories, checkedConditions, priceRange, sortBy]);
+  }, [checkedCategories, checkedConditions, priceRange, sortBy, listings]);
 
   const handleToggleCategory = (category) => () => {
     setCheckedCategories(prev => {
@@ -207,7 +203,7 @@ const HomePage = () => {
                     </Paper>
                   </Grid>
                   <Grid item xs={12}>
-                    <Typography variant="h4">Listings</Typography>
+                    <Typography variant="h4">{title}</Typography>
                     {filteredListings.map((listing) => (
                       <div
                         key={listing.id}
@@ -236,4 +232,4 @@ const HomePage = () => {
   );
 };
 
-export default HomePage;
+export default ListingsLayout;
