+++ src/Front-end/NavBar.js
@@ -21,4 +21,4 @@ const NavBar = () => {
     )
 }
 
- export default NavBar
+export default NavBar
