--- src/Watchlist.js
@@ -1,12 +1,0 @@
-import NavBar from "./NavBar";
-
-const Watchlist = () => {
-    return (
-        <>
-            <NavBar />
-            <h1>Watchlist</h1>
-        </>
-    )
-}
-
- export default Watchlist
